4:56

+++ src/main/java/com/Controllers/FoodController.java
@@ -47,10 +47,18 @@ public class FoodController extends HttpServlet {
             request.setAttribute("listCate", cateDAO.getAllCate());
             request.getRequestDispatcher("/listfood.jsp").forward(request, response);
         } else if (path.startsWith("/food/add")) {
+            if (session.getAttribute("admin") == null) {
+                response.sendRedirect("/home");
+                return;
+            }
             session.setAttribute("add_update", "add");
             session.setAttribute("FOOD", null);
             request.getRequestDispatcher("/AddNewFood.jsp").forward(request, response);
         } else if (path.startsWith("/food/update/")) {
+            if (session.getAttribute("admin") == null) {
+                response.sendRedirect("/home");
+                return;
+            }
             String[] s = path.split("/");
             String id = s[s.length - 1];
             FoodDAO dao = new FoodDAO();
@@ -63,11 +71,15 @@ public class FoodController extends HttpServlet {
                 request.getRequestDispatcher("/AddNewFood.jsp").forward(request, response);
             }
         } else if (path.startsWith("/food/delete/")) {
+            if (session.getAttribute("admin") == null) {
+                response.sendRedirect("/home");
+                return;
+            }
             String[] s = path.split("/");
             String id = s[s.length - 1];
             FoodDAO dao = new FoodDAO();
             dao.deleteFood(id);
-            response.sendRedirect("/food/list");
+            response.sendRedirect("/admin/foodmanager");
         } else if (path.startsWith("/food/false-")) {
             if (session.getAttribute("admin") == null) {
                 response.sendRedirect("/home");
@@ -88,6 +100,8 @@ public class FoodController extends HttpServlet {
             FoodDAO dao = new FoodDAO();
             dao.setFoodStatus(id, "true");
             response.sendRedirect("/admin/foodmanager");
+        } else {
+            response.sendRedirect("/error");
         }
         // code thêm cái sua xoa o day
     }
@@ -106,9 +120,9 @@ public class FoodController extends HttpServlet {
         HttpSession session = request.getSession();
         if (request.getParameter("btnAddNewFood") != null) {
             if (session.getAttribute("add_update").equals("add")) {
-                foodPost("/food/list", "/food/add", "add", request, response);
+                foodPost("/admin/foodmanager", "/food/add", "add", request, response);
             } else if (session.getAttribute("add_update").equals("update")) {
-                foodPost("/food/list", "/food/update", "update", request, response);
+                foodPost("/admin/foodmanager", "/food/update", "update", request, response);
             }
         }
     }
